chore(tts): move E2E TTS automation helper into __automation__
Keeps E2E-only orchestration out of utils so it can't be pulled into a production build.
